+++ android/app/src/main/java/com/smart/pantry/MainActivity.java
@@ -1,6 +1,12 @@
 package com.smart.pantry;
 
+import android.os.Bundle;
 import com.getcapacitor.BridgeActivity;
 
 public class MainActivity extends BridgeActivity {
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        registerPlugin(CookingModePlugin.class);
+        super.onCreate(savedInstanceState);
+    }
 }