+++ firebaseConfig.ts
@@ -1,6 +1,6 @@
 import { initializeApp } from "firebase/app";
 import { getAuth, setPersistence, browserLocalPersistence, indexedDBLocalPersistence } from "firebase/auth";
-import { getFirestore } from "firebase/firestore";
+import { initializeFirestore, persistentLocalCache, persistentMultipleTabManager } from "firebase/firestore";
 import { getStorage } from "firebase/storage";
 import type { Analytics } from "firebase/analytics";
 import { getAnalytics } from "firebase/analytics";
@@ -26,7 +26,11 @@ export const auth = getAuth(app);
 // auth.settings = {
 //   appVerificationDisabledForTesting: true
 // };
-export const db = getFirestore(app);
+export const db = initializeFirestore(app, {
+  localCache: persistentLocalCache({
+    tabManager: persistentMultipleTabManager(),
+  }),
+});
 export const storage = getStorage(app);
 export const functions = getFunctions(app);